requirements.txt
streamlit
pandas
gspread
google-auth
google-auth-oauthlib
google-api-python-client
Pillow
urllib3
